Sessions/Forth_Month/14-3-2026/Assignments/Cat_And_Dogs/requirements.txt.txt
TensorFlow==2.19.0
Keras==3.13.2
NumPy==2.0.2
Pillow==11.3.0
Gradio==5.50.0